add cluster set selection
