+++ frontend/src/routes/Applications/CreateApplication/Subscription/templates/templatePlacement.hbs
@@ -59,7 +59,9 @@ spec:
             {{/if}}
             {{/each}} 
   clusterSets:
-    - default
+    {{#each clusterSelector.clusterSetsList}}
+    - {{{clusterSetName}}}
+    {{/each}}
 {{else}}
   {{#if online-cluster-only-checkbox}}
 spec:
@@ -74,13 +76,16 @@ spec:
       'local-cluster': 'true'
   {{/if}}
 {{/if}}
-
+{{#each clusterSelector.clusterSetsList}}
+{{#unless existManagedClusterSetBinding}} 
 ---
 apiVersion: cluster.open-cluster-management.io/v1beta2
 kind: ManagedClusterSetBinding
 metadata:
   namespace: {{{@root.namespace}}}
-  name: default
+  name: {{{clusterSetName}}}
 spec:
-  clusterSet: default
+  clusterSet: {{{clusterSetName}}}
+{{/unless}}  
+{{/each}}
 {{/unless}}
