+++ frontend/src/routes/Infrastructure/Clusters/ManagedClusters/CreateCluster/components/assisted-installer/utils.ts
@@ -747,9 +747,10 @@ export const importYaml = (yamlContent: unknown) => {
 
 // Simple string search is _so far_ enough.
 // Full key-path support would be ineffective when not actually needed.
-export const getTemplateValue = (yaml: string, simpleKey: string, defaultValue: string) => {
+export const getTemplateValue = (yaml: string, simpleKey: string, defaultValue: string, index?: number) => {
   const lines = yaml.split('\n')
   const regex = new RegExp(`^ *${simpleKey}: *`)
+  let value
 
   const rows = lines.filter((l) => l.match(regex))
 
@@ -757,12 +758,17 @@ export const getTemplateValue = (yaml: string, simpleKey: string, defaultValue: 
     return defaultValue
   }
 
-  if (rows.length > 1) {
+  if (rows.length > 1 && index === undefined) {
     // Provide better key (i.e. leverage indentation). If this is not enough, let's full-parse the yaml instead.
     throw new Error(`Multiple matches for yaml key "${simpleKey}"`)
   }
 
-  const value = rows[0].replace(regex, '').trim()
+  if (rows.length > 1 && index) {
+    value = rows[index].replace(regex, '').trim()
+    return value
+  }
+
+  value = rows[0].replace(regex, '').trim()
   return value
 }
